Smoke test updates: Added flags test, updated makefiles and check_smoke to accomdate new test, added nvidia support for threads, aomp_mappings, flags

+++ test/smoke/check_smoke.sh
@@ -4,6 +4,14 @@
 # Tests that need to be visually inspected: devices, pfspecify, pfspecify_str, stream
 #
 #
+
+#Text Colors
+RED="\033[0;31m"
+GRN="\033[0;32m"
+BLU="\033[0;34m"
+ORG="\033[0;33m"
+BLK="\033[0m"
+
 path=$(pwd)
 count=0
 allTests=""
@@ -13,11 +21,12 @@ tempDir=""
 #Clean all testing directories
 make clean
 rm passing-tests.txt
+rm failing-tests.txt
 rm check-smoke.txt
 rm make-fail.txt
 
 echo ""
-echo "RUNNING ALL TESTS IN: $path "
+echo -e "$ORG"RUNNING ALL TESTS IN: $path"$BLK"
 echo ""
 
 echo "************************************************************************************" > check-smoke.txt
@@ -27,26 +36,71 @@ echo "**************************************************************************
 
 #Loop over all directories and make run / make check depending on directory name
 for directory in ./*/; do
-        shopt -s lastpipe
-        echo -n "$directory " | tr -d ./ | read tempDir
-        tempDir+=" "
-        allTests+="$tempDir"
+	shopt -s lastpipe
+	echo -n "$directory " | tr -d ./ | read tempDir
+	tempDir+=" "
+	allTests+="$tempDir"
 	let count++
 	(cd "$directory" && path=$(pwd) && base=$(basename $path) 
 		if [ $base == 'devices' ] || [ $base == 'pfspecifier' ] || [ $base == 'pfspecifier_str' ] || [ $base == 'stream' ] ; then 
 			make
-                        if [ $? -ne 0 ]; then
-                                echo "$base: Make Failed" >> ../make-fail.txt
-                        fi
+      if [ $? -ne 0 ]; then
+			echo "$base: Make Failed" >> ../make-fail.txt
+      fi
 			make run > /dev/null 2>&1
-		        make check > /dev/null 2>&1
+		  make check > /dev/null 2>&1
+
+		#flags has multiple runs
+		elif [ $base == 'flags' ] ; then
+			file="options.txt"
+
+			#Regex to search for "march=gfxXXX or sm_XX"
+			march_regex="(march=[a-z]+[0-9]*_?[0-9]+)"
+
+			#Read file and replace march with correct GPU and keep track of execution number
+			test_num=0;
+			while read -r line; do
+				((test_num++))
+				#if GPU is involved
+				if [[ "$line" =~ $march_regex ]]; then
+					march_match=${BASH_REMATCH[1]}
+					#remove march from command and replace with correct version
+					temp_line=${line/"-$march_match"}
+					mygpu=$AOMP_GPU
+					#If NVIDIA system, add nvptx targets, cuda, and remove amdgcn targets. This is done for testing purpose to avoid rewriting original amd command.
+					if [[ "$AOMP_GPU" == *"sm"* ]]; then
+						target_regex="(-fopenmp-[a-z]*=[a-z,-]*).*(-Xopenmp-[a-z]*=[a-z,-]*)"
+						if [[ "$line" =~ $target_regex ]]; then
+							target_match="${BASH_REMATCH[1]} ${BASH_REMATCH[2]}"
+							temp_line=${temp_line/"$target_match"}
+							nvidia_args="-fopenmp-targets=nvptx64-nvidia-cuda -Xopenmp-target=nvptx64-nvidia-cuda"
+							cuda_args="-L/usr/local/cuda/targets/x86_64-linux/lib -lcudart"
+						fi
+				  fi
+					#send variables to make
+					make make_options="$temp_line" nvidia_targets="$nvidia_args" march="-march=$mygpu" cuda="$cuda_args"
+					make make_options="$temp_line" nvidia_targets="$nvidia_args" march="-march=$mygpu" cuda="$cuda_args" test_num=$test_num check > /dev/null 2>&1
+					if [ $? -ne 0 ]; then
+						echo "$base $test_num: Make Failed" >> ../make-fail.txt
+					fi
+				else
+					make make_options="$line"
+					make make_options="$line" test_num=$test_num check > /dev/null 2>&1
+					if [ $? -ne 0 ]; then
+						echo "$base $test_num: Make Failed" >> ../make-fail.txt
+					fi
+				fi
+				echo ""
+			done < "$file"
+
 		else
 			make
-                        if [ $? -ne 0 ]; then
-                                echo "$base: Make Failed" >> ../make-fail.txt
-                        fi
+      if [ $? -ne 0 ]; then
+				echo "$base: Make Failed" >> ../make-fail.txt
+      fi
 			make check > /dev/null 2>&1
 		fi
+		echo ""
 	)
 	
 done
@@ -56,12 +110,12 @@ for directory in ./*/; do
 	(cd "$directory" && path=$(pwd) && base=$(basename $path)
 		if [ $base == 'devices' ] || [ $base == 'pfspecifier' ] || [ $base == 'pfspecifier_str' ] || [ $base == 'stream' ] ; then 
 			echo ""
-			echo "$base - Run Log:"
+			echo -e "$ORG"$base - Run Log:"$BLK"
 			echo "--------------------------"
 			cat run.log
 			echo ""
-		        echo ""	
-                fi
+		  echo ""
+    fi
 	)
 done
 
@@ -72,41 +126,55 @@ cat check-smoke.txt
 cat make-fail.txt
 echo ""
 
-#Clean all testing directories, hide output
-make clean > /dev/null 2>&1
+#Gather Test Data
+((total_tests=$(wc -l <  passing-tests.txt)))
+if [ -e make-fail.txt ]; then
+	((total_tests+=$(wc -l <  make-fail.txt)))
+fi
+if [ -e failing-tests.txt ]; then
+	((total_tests+=$(wc -l <  failing-tests.txt)))
+fi
 
-echo "-------------------- Results --------------------"
-echo "Number of tests: "$count
+#Print Results
+echo -e "$BLU"-------------------- Results --------------------"$BLK"
+echo -e "$BLU"Number of tests: $total_tests"$BLK"
 echo ""
-echo "Number of passing tests: " `wc -l <  passing-tests.txt`"/"$count
+echo -e "$GRN"Passing tests: `wc -l <  passing-tests.txt`/$total_tests"$BLK"
 echo ""
-
-#If the test is not found in the passing-tests file, add test to failedTests
-for test in $allTests
-do
-        if [ "$(grep -cx "$test" passing-tests.txt)" -eq 0 ];
-        then
-                failedTests+="$test "
-        fi
-done
 
 #Print failed tests
-echo "Failed Tests:"
-for fail in $failedTests
-do
-        echo "$fail"
-done
+echo -e "$RED"
+if [ -e failing-tests.txt ]; then
+	echo "Runtime Fails"
+	echo "--------------------"
+	cat failing-tests.txt
+	echo ""
+fi
+
+if [ -e make-fail.txt ]; then
+	echo "Compile Fails"
+	echo "--------------------"
+	cat make-fail.txt
+fi
+echo -e "$BLK"
 
 #Tests that need visual inspection
 echo ""
+echo -e "$ORG"
 echo "---------- Please inspect the output above to verify the following tests ----------"
 echo "devices"
 echo "pfspecifier"
 echo "pfspecifier_str"
 echo "stream"
+echo -e "$BLK"
 
 #Clean up, hide output
 make clean > /dev/null 2>&1
 rm check-smoke.txt
 rm passing-tests.txt
-rm make-fail.txt
+if [ -e failing-tests.txt ]; then
+	rm failing-tests.txt
+fi
+if [ -e make-fail.txt ]; then
+	rm make-fail.txt
+fi
